model_server/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
onnxruntime
numpy
tensorflow
pillow
python-multipart